+++ packages/algob/src/lib/status.ts
@@ -3,7 +3,7 @@ import type { modelsv2 } from "algosdk";
 import { AccountAddress, Deployer, Key, StateValue } from "../types";
 
 /**
- * Returns `account` asset holding of `assetID`. Returns undefined if the account is not
+ * Returns `account` balance of `assetID`. Returns 0 if the account has not
  * opt-in to the given asset id.
  * @param deployer algob deployer
  * @param accountAddress account to return assetholding info
@@ -13,15 +13,14 @@ export async function balanceOf (
   deployer: Deployer,
   accountAddress: AccountAddress,
   assetID: number
-): Promise<modelsv2.AssetHolding | undefined> {
+): Promise<number|bigint> {
   const accountInfo = await deployer.algodClient.accountInformation(accountAddress).do();
   for (const asset of accountInfo.assets) {
     if (asset['asset-id'] === assetID) {
-      console.log("Asset Holding Info:", asset);
-      return asset;
+      return asset.amount;
     }
   }
-  return undefined;
+  return 0n;
 };
 
 /**